Convert IRawElementProviderSimple to Cswin32 (#10177)
* Convert IRawElementProviderSimple to Cswin32

* Add tests and fix XAccessibleObjectTests

* more test updates

* Use Cswin32 definition of ProviderOptions

* Remove s_notificationEventAvailable from AccessibleObject and only use OS check

* Preserve sig and update Interop.AccessibleObject tests

* Place 1709 OSCheck inside PInvoke

* Remove nint -> VARIANT operator and return VT_I4 VARIANTs for handles

* use IUnknown* -> VARIANT explicit operator

* Add VARIANT.True and VARIANT.False

* Add WindowHandleToVariant helper method

* address other feedback

* remove unused variable

* remove unused using

* Update VARIANT.True and VARIANT.False

* Update new file TreeViewLabelEditAccessibleObjectTests

* clean up

* Add more detail doc and add debug assert

* Remove locale from link

+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/AccessibleObjects/ListViewItem.ListViewItemImageAccessibleObjectTests.cs
@@ -54,7 +54,7 @@ public class ListViewItem_ListViewItemImageAccessibleObjectTests
 
         AccessibleObject imageAccessibleObject = listViewItem.AccessibilityObject.GetChild(0);
 
-        Assert.Equal(UIA_CONTROLTYPE_ID.UIA_ImageControlTypeId, (UIA_CONTROLTYPE_ID)imageAccessibleObject.GetPropertyValue(UIA_PROPERTY_ID.UIA_ControlTypePropertyId));
+        Assert.Equal(UIA_CONTROLTYPE_ID.UIA_ImageControlTypeId, (UIA_CONTROLTYPE_ID)(int)imageAccessibleObject.GetPropertyValue(UIA_PROPERTY_ID.UIA_ControlTypePropertyId));
         Assert.False((bool)imageAccessibleObject.GetPropertyValue(UIA_PROPERTY_ID.UIA_HasKeyboardFocusPropertyId));
         Assert.False((bool)imageAccessibleObject.GetPropertyValue(UIA_PROPERTY_ID.UIA_IsKeyboardFocusablePropertyId));
     }
